requirements.txt
# 项目依赖包

# 深度学习框架
pytorch>=2.0.0

# 自然语言处理
jieba>=0.42.1

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 评估工具
scikit-learn>=1.3.0
matplotlib>=3.7.0

# Web服务
flask>=2.3.0
flask-cors>=4.0.0

# 其他
tqdm>=4.65.0    # 进度条

# 可选依赖
# 用于GPU加速（如果需要）
# cudatoolkit>=11.7
